+++ test-tsd/types.test-d.ts
@@ -1,0 +1,2 @@
+import { expectType } from 'tsd';
+import * as knex from '../types'